backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
mysqlclient==2.2.7
